[Day6] Complete
Still working on day5 part2, it's going to requier some more brute force or some elegant math.
Day 6 was fairly easy with Python. Puzzle2 was really about BigInts which are default in
Pythons implimentation so not really a problem.  Would have need a libarary for this in C/C++
or needed to use floats and hope that rounding errros were not an issue...